+++ capstone/src/main/webapp/business.html
@@ -14,7 +14,7 @@
       <ul class="nav">
         <li><a href="index.html">Home</a></li>
         <li><a href="profile.html">Profile</a></li>
-        <li id="logout-button">Log out</li>
+        <li id="auth-button"></li>
       </ul>
     </nav>
