+++ capstone/src/main/webapp/index.html
@@ -5,9 +5,8 @@
     <title>Home</title>
     <link rel="stylesheet" href="css/index.css">
     <script type="module" src="js/index.js"></script>
-    <script src="js/authentication.js"></script>
   </head>
-  <body onload="getLoginStatus()">
+  <body>
     <nav>
       <div class="logo">
         <h1>Support Local Businesses</h1>
