+++ capstone/src/main/webapp/profile.html
@@ -6,7 +6,7 @@
     <title>Profile</title>
     <link rel="stylesheet" href="css/style.css">
     <link rel="stylesheet" href="css/profile.css">
-    <script src="js/profile.js"></script>
+    <script type="module" src="js/profile.js"></script>
   </head>
   <body>
     <nav>
@@ -16,7 +16,7 @@
       <ul class="nav">
         <li><a href="index.html">Home</a></li>
         <li><a href="profile.html">Profile</a></li>
-        <li id="logout-button">Log out</li>
+        <li id="auth-button"></li>
       </ul>
     </nav>
