+++ capstone/src/main/webapp/questionnaire.html
@@ -7,18 +7,8 @@
     <link rel="stylesheet" href="css/style.css">
   </head>
   <body>
-    <nav>
-      <div class="logo">
-        <h1>Support Local Businesses</h1>
-      </div>
-      <ul class="nav">
-        <li><a href="index.html">Home</a></li>
-        <li><a href="profile.html">Profile</a></li>
-        <li id="logout-button">Log out</li>
-      </ul>
-    </nav>
-
     <div class="content">
+      <!-- TODO: Edit this HTML page with a form. -->
       <h1>Are you a business owner?</h1>
       <p> Yes </p>
       <p> No </p>
